select and spatial_selection modules merged in filter
all the spatial filtering functions are now available with select_polygon
